defined all the different gui modes defining user writes and gen vars for different things to be tested with a human. compiles but still need to test

+++ MIT_DLEG/inc/machine_learning_methods.h
@@ -116,7 +116,7 @@ struct statistics_s
 	float* x;
 	float* y;
 
-	uint8_t k_est;
+	int k_est;
 
 
 	//Segmentation, state,and mutex variables
